LiquidAssociation created

+++ contracts/IToken.sol
@@ -1,5 +1,5 @@
 pragma solidity ^0.4.2;
 
 contract IToken {
-    mapping (address => uint256) public balanceOf;
+    mapping (address => uint) public balanceOf;
 }
